필터링 은행 체크 리스트 context로 전역 관리 -> 모달이 다시 열릴 때, 체크 상태 유지
